moved Sponsors elements in the right folder
