+++ src/components/Home/Home.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import WelcomeCarousel from './WelcomeCarousel';
 import Banner from './Banner';
-// import Translation from './Translation';
-import Sponsors from './Sponsors';
+import Sponsors from '../Sponsors/Sponsors';
 import Review from './Review';
 
 const Home = () => {
