+++ src/components/Sponsors/Sponsors.scss
